refactor(admin): split qa review route
Why: keep the Quantercise QA page focused on route and Suspense wiring while moving the remote QA load and review UI into named modules.

Verify: pnpm exec prettier --check apps/admin/src/app/quantercise/qa/page.tsx apps/admin/src/app/quantercise/qa/qa-content.tsx apps/admin/src/app/quantercise/qa/qa-sections.tsx; git diff --check -- apps/admin/src/app/quantercise/qa/page.tsx apps/admin/src/app/quantercise/qa/qa-content.tsx apps/admin/src/app/quantercise/qa/qa-sections.tsx; pnpm --filter @anipotts/admin typecheck; pnpm --filter @anipotts/admin lint; pnpm --filter @anipotts/admin build; pnpm validate
